+++ io/CHANGES.md
@@ -1,5 +1,9 @@
 # unreleased 0.6.0
+## Add
+- add `io::{AsyncBufWrite, AsyncBufRead}` impl for `net::{TcpStream, UnixStream}`
+
 ## Change
+- `io::AsyncBufWrite::shutdown` becomes async method 
 - update `tokio-uring-xitca` to `0.2.0`
 
 # 0.5.1